@@ -1,5 +1,5 @@
 class CustomPagesController < ApplicationController
-  skip_before_filter :show_global_container
+  skip_before_action :show_global_container
 
   def how_to_master; end
 
