@@ -2,7 +2,7 @@ module Billing::Subscriptionable
   extend ActiveSupport::Concern
 
   included do
-    scope :active, -> { where{ active_to > DateTime.current && active_from < DateTime.current } }
+    scope :active, -> { where.has{ active_to > DateTime.current && active_from < DateTime.current } }
 
     validates :active_to, :active_from, presence: true
   end
