@@ -15,7 +15,7 @@ class PlaceSearch < BaseSearch
     def searchkick_search
       Searchkick.search query,
         index_name: Place,
-        fields: ['name^10', 'place_type_name^10', 'full_name^2', 'search_string'],
+        fields: %w[name^10 place_type_name^10 full_name^2 search_string],
         match: :word_start,
         page: page,
         per_page: per_page,
