@@ -27,7 +27,7 @@ describe Admin::PagesController, type: :controller do
 
   describe 'GET #show' do
     before :each do
-      get :show, id: page.id
+      get :show, params: { id: page.id }
     end
 
     it 'checks the status' do
@@ -48,14 +48,14 @@ describe Admin::PagesController, type: :controller do
 
   describe 'POST #create' do
     it 'checks the redirect status' do
-      post :create, page: params
+      post :create, params: { page: params }
       expect(response.status).to eq(302)
     end
   end
 
   describe 'GET #edit' do
     before :each do
-      get :edit, id: page.id
+      get :edit, params: { id: page.id }
     end
 
     it 'checks the status' do
@@ -69,7 +69,7 @@ describe Admin::PagesController, type: :controller do
 
   describe 'PATCH #update' do
     before :each do
-      patch :update, id: page.id, page: params
+      patch :update, params: { id: page.id, page: params }
     end
 
     it 'checks the redirect status' do
@@ -79,7 +79,7 @@ describe Admin::PagesController, type: :controller do
 
   describe 'DELETE #destroy' do
     it 'checks the status' do
-      delete :destroy, id: page.id
+      delete :destroy, params: { id: page.id }
       expect(response.status).to eq(302)
     end
   end
